requirements.txt
flask
Flask-SQLAlchemy
flask-login
Flask-Security
flask_cors
flask_admin